Audio: sanitize `volume`
* move bound checkes to platform-independent layer
* snap to 0 / 1 to use memcpy/memset when close to those factors
* use float instead of qreal in the private volume APIs consistently
* use epsilon sufficient for 22bit resolution

Pick-to: 6.8 6.9

+++ src/multimedia/android/qandroidaudiosource.cpp
@@ -441,12 +441,12 @@ qint64 QAndroidAudioSource::processedUSecs() const
     return m_format.durationForBytes(m_processedBytes);
 }
 
-void QAndroidAudioSource::setVolume(qreal vol)
+void QAndroidAudioSource::setVolume(float vol)
 {
     m_volume = vol;
 }
 
-qreal QAndroidAudioSource::volume() const
+float QAndroidAudioSource::volume() const
 {
     return m_volume;
 }
